Stop uWSGI from trying to load an invalid vassal
We only have two vassals, uwsgi-halliganhelper.ini and
uwsgi-halliganhelper-sockets.ini. However, we have a third
file in the vassals directory, uwsgi-base.ini, which sets
environment variables common to both "real" vassals. uWSGI
was previously attempting to load this file as a separate
vassal every two minutes and failing, causing 27 lines to
be logged by systemd, which quite effectively obscured any
real issues in the logs.

Rename the base file to have an extension that uWSGI doesn't
attempt to load automatically.